--- visionhub-cli/src/config-template.yaml
@@ -1,93 +1,0 @@
-slug: 
-  description: uniq name of the model
-  type: str
-name: 
-  description: Descriptive name of the model
-  type: str
-anons: 
-  description: Short description of the model
-  type: str
-description: 
-  description: Long description of the model
-  type: str
-md_documentation: 
-  description: path to md file with documentation 
-  default: README.md
-  type: path
-version: 
-  description: Version of the model
-  type: str
-link: 
-  description: Address of docker image
-  type: str
-modes: 
-  description: multiple of [IMG2IMG, IMG2VID, VID2IMG, VID2VID], separated by comma
-  type: array[str]
-preview: 
-  description: path to image
-  type: path
-image_input_example: 
-  description: Image input example
-  type: path
-  required: False
-video_input_example: 
-  description: Video input example
-  type: path
-  required: False
-image_output_example: 
-  description: Image output example
-  type: path
-  required: False
-video_output_example: 
-  description: Video output example
-  type: path
-  required: False
-cost_for_image_to_image: 
-  description: Cost of one image
-  type: float
-  required: False
-  default: 0
-cost_for_image_to_video: 
-  description: Cost of one image that will converted to video
-  type: float
-  required: False
-  default: 0
-cost_for_video_to_image: 
-  description: Cost of one second of video that will converted to image
-  type: float
-  required: False
-  default: 0
-cost_for_video_to_video: 
-  description: Cost of one second of video
-  type: float
-  required: False
-  default: 0
-without_meta: 
-  description: Can be run without meta
-  type: boolean
-  default: y
-is_private: 
-  description: Is it private model
-  type: boolean
-gpu: 
-  description: Is model running on a gpu
-  type: boolean
-  default: n
-batch_size: 
-  description: Size of the batch
-  type: int
-is_video_stateful: 
-  description: Is video stateful
-  type: boolean
-meta_template: 
-  description: Json template for the meta
-  type: path | str
-  required: False
-meta_input_example: 
-  description: Json example of the meta
-  type: path | str
-  required: False
-prediction_example: 
-  description: Example of the prediction
-  type: path | str
-  required: False
